Corrective-RAG/requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-tavily
langgraph
chromadb
openai
pydantic
streamlit
